save json when you rate a book

+++ stefano_vecchiati_test/stefano_vecchiati_test/Services/Models/BookModel.swift
@@ -10,6 +10,6 @@ struct BookModel: Codable {
     var title: String?
     var description: String?
     var imageName: String?
-    var rate: Int? = 0
+    var rate: Double? = 0
     
 }
